feat(stock): สินค้าเงินเป็นล็อต — 1 เลขสต็อกถือได้หลายชิ้น
ระบบเดิมบังคับ 1 เลขสต็อก = 1 ชิ้น ทำให้ตัวโอนงานเงินแตก 1 รหัสเก่า
เป็นหลายเลขใหม่ (2,346 รหัส -> 12,941 เลข) ลูกค้าเห็นรหัสซ้ำกันเต็มหน้าจอ
และนับสต็อกจริงไม่ได้ จึงเปลี่ยนให้ piece ถือจำนวนได้เอง

- tbt_stock_piece เพิ่ม qty / qty_reserved (ทอง default 1/0 พฤติกรรมเดิมทุกอย่าง)
- StockPieceQtyHelper เป็นที่เดียวที่คำนวณ available และตัดสิน status
  (qty<=0 -> SOLD, qty_reserved>=qty -> RESERVED, อื่น ๆ IN_STOCK)
  ทำให้ filter pieceStatus=IN_STOCK เดิมยังแปลว่า "มีของพร้อมขาย" เหมือนเดิม
- ตัวโอนงานเงิน: 1 แถวคิว = 1 piece ที่ qty=N (เดิมสร้าง N piece)
- รับสินค้าแผนเงิน: auto-group แถวที่เหมือนกันเป็นล็อตเดียว ไม่ต้องเพิ่มช่องกรอก
- ขาย/จอง/คืน/POS: ตัดและคืนเป็นจำนวน จองเกินของพร้อมขายไม่ได้
- ย้ายคลัง: ย้ายทั้งล็อต และปฏิเสธเมื่อมีการจองค้าง (ย้ายบางส่วนไว้เฟส 2)
- reconciliation + รายงาน/dashboard: นับจาก SUM(qty) แทนการนับแถว
- API ส่ง qty/qtyReserved/qtyAvailable ของ piece แทนค่าคงที่ 1

migration 20260908_01 เพิ่มคอลัมน์ พร้อม backfill แถวที่ SOLD/RESERVED อยู่ก่อนแล้ว
(ไม่งั้นยกเลิกใบแจ้งหนี้เก่าจะทำ qty งอก และออกบิลจากของที่จองไว้จะทำยอดจองติดลบ)
migration 20260908_02 รวมล็อตงานเงินย้อนหลัง — นับเฉพาะชิ้นที่ยัง IN_STOCK
ปล่อยชิ้นที่ขายไปแล้วไว้ทั้งแถว ข้ามล็อตที่ขายหมด และ abort ตัวเองถ้ามีการจองค้าง
หรือมีเอกสารอ้างชิ้นที่กำลังจะถูกลบ

+++ Jewelry.Api/jewelry.Model/Stock/Reconciliation/ReconciliationReport.cs
@@ -21,6 +21,6 @@ namespace Jewelry.Model.Stock.Reconciliation
         public string SkuCode { get; set; } = null!;
         public string LocationCode { get; set; } = null!;
         public decimal BalanceQty { get; set; }
-        public int PieceCount { get; set; }
+        public decimal PieceCount { get; set; }
     }
 }
